refactor(jobpost): move to career app (#331)
* refactor(jobpost): move to career app

* format

* Update migrations
