+++ app/career/admin.py
@@ -1,5 +1,7 @@
 from django.contrib import admin
 
-from app.career import models
+from app.career.models.job_post import JobPost
+from app.career.models.weekly_business import WeeklyBusiness
 
-admin.site.register(models.WeeklyBusiness)
+admin.site.register(JobPost)
+admin.site.register(WeeklyBusiness)
