+++ app/content/filters/__init__.py
@@ -1,4 +1,3 @@
 from app.content.filters.cheatsheet import CheatsheetFilter
 from app.content.filters.event import EventFilter
-from app.content.filters.job_post import JobPostFilter
 from app.content.filters.user import UserFilter
