+++ app/content/models/__init__.py
@@ -2,7 +2,6 @@ from app.content.models.badge import Badge
 from app.content.models.category import Category
 from app.content.models.cheatsheet import Cheatsheet
 from app.content.models.event import Event
-from app.content.models.job_post import JobPost
 from app.content.models.news import News
 from app.content.models.user import User, UserManager
 from app.content.models.user_badge import UserBadge
