+++ app/content/urls.py
@@ -6,7 +6,6 @@ from app.content.views import (
     CategoryViewSet,
     CheatsheetViewSet,
     EventViewSet,
-    JobPostViewSet,
     NewsViewSet,
     NotificationViewSet,
     PageViewSet,
@@ -28,7 +27,6 @@ router.register("events", EventViewSet, basename="event")
 router.register("warning", WarningViewSet, basename="warning")
 router.register("category", CategoryViewSet)
 router.register("short-link", ShortLinkViewSet, basename="short-link")
-router.register("jobpost", JobPostViewSet, basename="jobpost")
 router.register("user", UserViewSet, basename="user")
 router.register(
     r"events/(?P<event_id>\d+)/users", RegistrationViewSet, basename="registration"
